Fixed fxaa shaders to initialize integers

+++ shaders/fxaa_v.glsl
@@ -1,8 +1,9 @@
+//GLSL
 #version 120
 varying vec4 posPos;
 uniform float FXAA_SUBPIX_SHIFT = 1.0/4.0;
-uniform float rt_w;
-uniform float rt_h;
+uniform int rt_w;
+uniform int rt_h;
 void main()
     {
     gl_Position = ftransform();
